feat(app): fix some problems and add config tool and add setup&uninstall

+++ Project/MyProcControlLite/setupui.cpp
@@ -1,19 +1,39 @@
 ﻿#include "setupui.h"
 #include <w32use.hpp>
 #include <shlobj.h>
+#include "configtool.hpp"
+#include "resource.h"
 using namespace std;
 
 extern HINSTANCE hInst;
 
 
-int RunSetupUI(std::wstring name, std::wstring action) {
+int RunConfigUI(std::wstring name) {
+	INITCOMMONCONTROLSEX icce{};
+	icce.dwSize = sizeof(icce);
+	icce.dwICC = ICC_ALL_CLASSES;
+	InitCommonControlsEx(&icce);
+	bool withName = !name.empty();
+	if (!withName) name = L"MyProcControl-Lite";
+
+	MyProcControl_Lite::ConfigTool win(name, withName);
+	win.create();
+	win.set_main_window();
+	win.center();
+	win.show();
+	return win.run();
+}
+
+int RunSetupUI(std::wstring name, std::wstring action, bool interactive) {
+	if (action.empty()) return RunConfigUI(name);
 	INITCOMMONCONTROLSEX icce{};
 	icce.dwSize = sizeof(icce);
 	icce.dwICC = ICC_ALL_CLASSES;
 	InitCommonControlsEx(&icce);
 	if (name.empty()) name = L"MyProcControl-Lite";
+	int btn = 0;
 	if (action.empty()) {
-		int btn = 0;
+		if (!interactive) return ERROR_AMBIGUOUS_SYSTEM_DEVICE;
 		TaskDialog(NULL, hInst, L"MyProcControl (Lite) Setup", L"What would you want to do?",
 			L"Press [Yes] to Install.\n"
 			L"Press [No] to Uninstall.\n"
@@ -44,7 +64,8 @@ int RunSetupUI(std::wstring name, std::wstring action) {
 			auto program = make_unique<WCHAR[]>(32768);
 			GetModuleFileNameW(NULL, program.get(), 32768);
 			sei.lpFile = program.get();
-			wstring params = L"--type=setup --action=" + action;
+			wstring params = L"--type=setup --action=" + action + L" --name=\"" + name + L"\"";
+			if (interactive) params += L"  --interactive";
 			sei.lpParameters = params.c_str();
 			if (!ShellExecuteExW(&sei)) goto setup_askprivilege;
 			if (!sei.hProcess) return GetLastError();
@@ -56,6 +77,7 @@ int RunSetupUI(std::wstring name, std::wstring action) {
 			return exitCode;
 		}
 	} // if (action.empty())
+	if (!IsUserAnAdmin() && interactive) goto setup_askprivilege;
 	try {
 		ServiceManager scm;
 		if (action == L"install") {
@@ -85,12 +107,18 @@ int RunSetupUI(std::wstring name, std::wstring action) {
 			ChangeServiceConfig2W(myService, SERVICE_CONFIG_DESCRIPTION, &a);
 			if (!myService.start()) throw w32oop::exceptions::system_exception("Failed to start service.");
 
-			TaskDialog(NULL, NULL, L"MyProcControl (Lite) Setup",
+			if (interactive) TaskDialog(NULL, NULL, L"MyProcControl (Lite) Setup",
 				L"The product has been successfully installed to your computer.",
 				L"Click OK to dismiss.", TDCBF_OK_BUTTON | TDCBF_CANCEL_BUTTON, TD_INFORMATION_ICON, NULL);
 			return 0;
 		}
 		if (action == L"uninstall") {
+			if (interactive) {
+				TaskDialog(NULL, hInst, L"Uninstall MyProcControl (Lite)",
+					L"Are you sure you want to uninstall this product?", name.c_str(),
+					TDCBF_YES_BUTTON | TDCBF_CANCEL_BUTTON, MAKEINTRESOURCEW(IDI_ICON_APP), &btn);
+				if (btn != IDYES) return ERROR_CANCELLED;
+			}
 			Service myService = scm.get(name);
 			if (!myService.remove()) throw w32oop::exceptions::system_exception("Failed to remove service.");
 			if (myService.status() == SERVICE_RUNNING) {
@@ -100,22 +128,25 @@ int RunSetupUI(std::wstring name, std::wstring action) {
 			if (myService.status() == SERVICE_PAUSED) {
 				if (!myService.stop()) throw w32oop::exceptions::system_exception("Failed to stop service.");
 			}
-
-			TaskDialog(NULL, NULL, L"MyProcControl (Lite) Setup",
+			RegistryKey uninstall(HKEY_LOCAL_MACHINE, L"SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Uninstall");
+			uninstall.delete_key(L"Service_" + name);
+				
+			if (interactive) TaskDialog(NULL, NULL, L"MyProcControl (Lite) Setup",
 				L"The product has been successfully removed from your computer.",
 				L"Click OK to dismiss.",
 				TDCBF_OK_BUTTON | TDCBF_CANCEL_BUTTON, TD_INFORMATION_ICON, NULL);
 			return 0;
 		}
-		MessageBoxW(NULL, L"Unknown action type", L"Error", MB_ICONERROR);
+		if (interactive) MessageBoxW(NULL, L"Unknown action type", L"Error", MB_ICONERROR);
 		return 87;
 	}
 	catch (w32oop::exceptions::system_exception& e) {
-		TaskDialog(NULL, nullptr, L"MyProcControl (Lite) Setup",
+		if (interactive) TaskDialog(NULL, nullptr, L"MyProcControl (Lite) Setup",
 			L"The operation failed for an operation-specific reason",
 			(w32oop::util::str::converts::str_wstr(e.what()) + L"\n\n" +
 				ErrorChecker().message()).c_str(), TDCBF_CANCEL_BUTTON, TD_ERROR_ICON, 0);
-		return GetLastError();
+		DWORD err = GetLastError();
+		return int((!!err) ? err : 1);
 	}
 }
